+++ projects/ngx-material/src/components/index.ts
@@ -3,3 +3,4 @@ export * from './loading-overlay/loading-overlay.component';
 export * from './header/header.component';
 export * from './divider/ngx-divider.component';
 export * from './icon-button/icon-button.component';
+export * from './icon-anchor/icon-anchor.component';